// src/other-projects/analitycs-consumer/src/modules/analytics/conversation-analytics/interfaces/metrics.interface.ts
export interface MetricsInterface {
    metricsAssignmentAt: number;
    metricsCloseAt: number;
    metricsLastAgentReplyAt: number;
    metricsLastUserReplyAt: number;
    metricsMedianTimeToAgentReply: number;
    metricsMedianTimeToUserReply: number;
    metricsTimeToAgentReply: number;
    metricsTimeToAssignment: number;
    metricsTimeToClose: number;
    metricsFirstAgentReplyAt: number;
    metricsTimeToUserReply: number;
    metricsAwaitingWorkingTime?: number;
    metricsAutomaticDurationAttendance?: number;
    suspendedUntil: number;
    waitingSince: number;
    order: number;
    priority: number;
    whatsappExpiration: number;
}